Change how JSON is delivered in the user API
The User API currently sends JSON payload with an text/html header. This leads to JSON being displayed as if it was HTML, with strange results.

The correct Content-Type according to RFC 4627 is application/json. When sending the appropriate Content-Type, there is no need to further escape the JSON payload.

Finally, add a newline after the JSON data, so it doesn't screw up the terminal when you use curl.