Revert "Update translations"
This reverts commit b6347220ae4c41b6f719e5097f51ebfb64de63ec.

The current translation workflow will pull the .pot files into
Transifex where people will manage the translations. If we correct a
misspelled word it will be reflected in a new source string so people
will just need to do the translation again. While correcting .po files
may be useful to correct a running service, doing so at this stage seems
a waste of time.

Co-authored-by: Tomasz Wolniewicz <twoln@umk.pl>